Move punchout POST to front
